Differentiate the custom home page content
